StupidRobot: Fix cut to end-screen after 7 words, re_init word limit
This commit fixes the StupidRobot code intended to cut to the end-screen early
if the player enters seven accepted words. Previously, if the user entered all
seven words, they would have to wait for the timer to run out or click "Quit"
to continue.

Specifically, it fixes a conditional (acceptedWords > maxLevel) to be
(acceptedWords > maxLevel - 4) because the user cannot enter 1, 2, or 3-letter
words.

Other changes:
 - Changed definition of maxLevel in re_init to 10.
   The re_init function is intended to reset various game constants to their
   initial values. However, the maxLevel in re_init was not changed from 13 to
   10 when the original constant was. This meant that in subsequent games,
   players could type longer words, although they would not be accepted.
   (This also caused problems with the conditional mentioned above.)