Improve displaySession parser, add new StupidRobot events
These changes reflect feedback on the original displaySession parser. The actor
for each event (player/game) has been capitalized for emphasis. Repeats of the
image being displayed have been removed. In addition, two additional event
types have been added to the log: "start session" and "end game". The first is
logged when the game starts, and the second when the score screen is displayed.