StupidRobot: Add in-game score display to final screen Adds an in-game score display to accompany the StupidRobot scoring algorithm. Note that the two are completely independent - while it is possible to falsify the in-game display, it will not affect the overall scoreboard.
As a part of adding the score display, the HTML structure of the final screen
has been redone, removing the questionable use of <span> elements and
line breaks, and replacing them with a table. This also allows for easy
alignment of the score counters with their respective words. In addition, the
empty 11- to 13-letter word spaces remaining from an earlier revision have now
been removed entirely.

Aesthetic changes will likely be made to the display before merging this branch
back into the development branch, but the functionality is finished. In
addition, there is a known bug where the display will sometimes not animate;
this is an inconsistent problem and the cause is not yet known.