StupidRobot: Fix flashMessage overwriting final screen
The flashMessage function is used to display StupidRobot's response to a user
submission, first showing the response, then restoring the text to its
original prompt ("Human: what is in this image?"). However, if it was called
and then the final screen was immediately shown (either by time limit or by
clicking the "Quit" button), then it would overwrite the final response ("I
now comprehend X more words...") with the original prompt.

This was due to an HTML element ID conflict between the in-game and final
screen messages. The latter's ID has been changed from "gameMessage" to
"endgameMessage", resolving the issue.