Fix GuessWhat to prevent multiword hint exploit
After the change to multiword parsing (commit 7453f3b9), it became much
easier for users to enter multiword hints. This has been fixed using
the same code that ZenTag uses, with some slight modifications. This
limits the characters that can be entered in the text area to just
letters (a-z, A-Z).